Subclass Plugin locators to be aware of extension gems (closes gh-135)
This creates a new plugin type: Radiant::ExtensionGem. It only differs from
normal GemPlugins in that it won't be treated as an engine, meaning its load
paths will be prioritized above the base application's. Also, the plugin
locator is split into two classes that are aware of the new plugin type.