move views, controllers, and helpers under the radiant/admin namespace where necessary
This is a large commit and includes some unrelated changes.
When trying to get the application working on the spec/dummy app, I made changes to the way some controller
methods are defined in the provided application controller.

This also adds an admin controller which provides the session and login methods.

A minor change to the login form page was made to remove reference to the type of boot (gem vs. vendor)